Revert "replaced mixin object with BaseCommand"
